+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * BraintreeDataCollector
   * Update PPRiskMagnes to 5.6.0 
   * Add `collectDeviceDataOnSuccess(riskCorrelationID: completion:)` method that ensures Magnes SDK completes data submission prior to returning device data.
+* BraintreePayPal
+    * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events.
 
 ## 7.3.0 (2026-01-29)
 * BraintreePayPal